requirements.txt
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
